requirements.txt
streamlit>=1.30.0
numpy<2.0.0
opencv-python-headless
easyocr
Pillow
pandas